feat: Extracted rectangle logic to separate module
