+++ src/shapes/rectangle/rectangle.h
@@ -1,0 +1,2 @@
+void rectangleArea();
+void rectanglePerimeter();
